Design : #24 내용으로 컴포넌트 디자인 변경 및 app 코드 삭제
